% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_NAsamples_log}
\alias{add_NAsamples_log}
\title{Add NA and then do log10 transformation}
\usage{
add_NAsamples_log(df, sampleTable)
}
\arguments{
\item{df}{}

\item{sampleTable}{}
}
\description{
Add NA and then do log10 transformation
}
