% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcoat}
\alias{bcoat}
\title{BCoAT tree related data}
\format{
list
A list with two values:
\describe{
  \item{tree}{Phylogenetic tree}
  \item{anno}{annotation of nodes in DataFrame}
  ...
}
}
\source{
{Created in-house data}
data(bcoat)
}
\usage{
bcoat
}
\description{
BCoAT tree related data
}
\keyword{datasets}
