% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_vc_by_quality}
\alias{extract_vc_by_quality}
\title{Extract VC by checkv quality}
\usage{
extract_vc_by_quality(df, quality)
}
\arguments{
\item{df}{}

\item{quality}{}
}
\description{
Extract VC by checkv quality
}
