% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvhdvirome-package.R
\docType{package}
\name{gvhdvirome-package}
\alias{gvhdvirome}
\alias{gvhdvirome-package}
\title{gvhdvirome: Code and analyses for the GVHD virome paper}
\description{
This package contains the code and analyses for the GVHD virome paper.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rujinlong/gvhdvirome}
  \item Report bugs at \url{https://github.com/rujinlong/gvhdvirome/issues}
}

}
\author{
\strong{Maintainer}: Jinlong Ru \email{jinlong.ru@gmail.com} (\href{https://orcid.org/0000-0002-6757-6018}{ORCID})

}
\keyword{internal}
