% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{merge_vOTU_abundance}
\alias{merge_vOTU_abundance}
\title{Merge vOTU abundance}
\usage{
merge_vOTU_abundance(abundance_contigs_count, vOTUs)
}
\arguments{
\item{abundance_contigs_count}{}

\item{vOTUs}{}
}
\description{
Merge vOTU abundance
}
