% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p0064gvhd-package.R
\docType{package}
\name{p0064gvhd-package}
\alias{p0064gvhd}
\alias{p0064gvhd-package}
\title{p0064gvhd: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rujinlong/p0064gvhd}
  \item Report bugs at \url{https://github.com/rujinlong/p0064gvhd/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
