% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_weights_fn}
\alias{plot_weights_fn}
\title{Plot MOFA weights}
\usage{
plot_weights_fn(mofa, factor = 1, view = 1, nfeatures = 10)
}
\arguments{
\item{mofa}{}

\item{factor}{}

\item{view}{}

\item{nfeatures}{}
}
\description{
Plot MOFA weights
}
