% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plt_ordination}
\alias{plt_ordination}
\title{Plot ordination using phyloseq object}
\usage{
plt_ordination(pseq, fout, method)
}
\arguments{
\item{pseq}{}

\item{fout}{}

\item{method}{}
}
\description{
Plot ordination using phyloseq object
}
