% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plt_richness}
\alias{plt_richness}
\title{Plot richness using phyloseq object}
\usage{
plt_richness(pseq, fout, dis_stat = dis_stat)
}
\arguments{
\item{pseq}{}

\item{fout}{}

\item{dis_stat}{}
}
\description{
Plot richness using phyloseq object
}
