% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_metabolites}
\alias{read_metabolites}
\title{Read metabolism data}
\usage{
read_metabolites(df_metabolites, metadata)
}
\arguments{
\item{df_metabolites}{}

\item{metadata}{}
}
\description{
Read metabolism data
}
