% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcontact2.R
\name{read_vConTACT2}
\alias{read_vConTACT2}
\title{Title}
\usage{
read_vConTACT2(fpath, assembler_label = "_NODE_")
}
\arguments{
\item{fpath}{Path of vConTACT2 genome cluster results}

\item{assembler_label}{"_NODE_" for SPAdes}
}
\value{
list of results
}
\description{
Title
}
