% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_physeq_metadata}
\alias{update_physeq_metadata}
\title{Update phyloseq metadata}
\usage{
update_physeq_metadata(pseq, fp_metadata, sample_order)
}
\arguments{
\item{pseq}{}

\item{fp_metadata}{}

\item{sample_order}{}
}
\description{
Update phyloseq metadata
}
