% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_coverm}
\alias{read_coverm}
\title{Read CoverM abundance file}
\usage{
read_coverm(fpath, cov = 0, fbin2contig = 0)
}
\arguments{
\item{fpath}{CoverM output file name}

\item{cov}{Coverage fraction file or not}

\item{fbin2contig}{Contig binning to contig mapping}
}
\value{
A dataframe
}
\description{
Read CoverM abundance file
}
