% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_vConTACT2}
\alias{read_vConTACT2}
\title{Parse vConTACT2 results}
\usage{
read_vConTACT2(fpath, assembler_label = "_NODE_")
}
\arguments{
\item{fpath}{Path of vConTACT2 genome cluster results}

\item{assembler_label}{"_NODE_" for SPAdes}
}
\value{
list of results
}
\description{
Parse vConTACT2 results
}
